BackUpThaventha upto TestGrouping save Success and TestCase Import success

+++ src/main/java/com/ii/testautomation/controllers/TestTypesController.java
@@ -7,6 +7,8 @@ import com.ii.testautomation.response.common.BaseResponse;
 import com.ii.testautomation.response.common.ContentResponse;
 import com.ii.testautomation.response.common.FileResponse;
 import com.ii.testautomation.response.common.PaginatedContentResponse;
+import com.ii.testautomation.service.ProjectService;
+import com.ii.testautomation.service.TestGroupingService;
 import com.ii.testautomation.service.TestTypesService;
 import com.ii.testautomation.utils.Constants;
 import com.ii.testautomation.utils.EndpointURI;
@@ -21,9 +23,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @RestController
 @CrossOrigin
@@ -32,8 +32,12 @@ public class TestTypesController {
     private TestTypesService testTypesService;
     @Autowired
     private StatusCodeBundle statusCodeBundle;
+    @Autowired
+    private TestGroupingService testGroupingService;
+    @Autowired
+    private ProjectService projectService;
 
-    @PostMapping(EndpointURI.TESTTYPE)
+    @PostMapping(EndpointURI.TEST_TYPE)
     public ResponseEntity<Object> insertTestTypes(@RequestBody TestTypesRequest testTypesRequest) {
         if (testTypesService.isExistsTestTypeByName(testTypesRequest.getName()))
             return ResponseEntity.ok(new BaseResponse(
@@ -46,10 +50,9 @@ public class TestTypesController {
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
                 statusCodeBundle.getInsertTestTypesSuccessMessage()));
-
     }
 
-    @PutMapping(EndpointURI.TESTTYPE)
+    @PutMapping(EndpointURI.TEST_TYPE)
     public ResponseEntity<Object> updateTestTypes(@RequestBody TestTypesRequest testTypesRequest) {
         if (!testTypesService.isExistsTestTypeById(testTypesRequest.getId()))
             return ResponseEntity.ok(new BaseResponse(
@@ -67,20 +70,26 @@ public class TestTypesController {
         return ResponseEntity.ok(new BaseResponse(
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
-                statusCodeBundle.getUpdateTestTypeSuccessMessage())); }
+                statusCodeBundle.getUpdateTestTypeSuccessMessage()));
+    }
 
-    @DeleteMapping(EndpointURI.TESTTYPE_BY_ID)
+    @DeleteMapping(EndpointURI.TEST_TYPE_BY_ID)
     public ResponseEntity<Object> deleteTestTypeById(@PathVariable Long id) {
         if (!testTypesService.isExistsTestTypeById(id))
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getTestTypeNotExistCode(),
                     statusCodeBundle.getTestTypeIdNotFoundMessage()));
+        if (testGroupingService.existsByTestTypesId(id))
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
+                    statusCodeBundle.getTestTypeDependentCode(),
+                    statusCodeBundle.getTestTypeDependentMessage()));
         testTypesService.deleteTestTypeById(id);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
-                statusCodeBundle.getDeleteTestTypesSuccessMessage()));}
+                statusCodeBundle.getDeleteTestTypesSuccessMessage()));
+    }
 
-    @GetMapping(EndpointURI.TESTTYPE_BY_ID)
+    @GetMapping(EndpointURI.TEST_TYPE_BY_ID)
     public ResponseEntity<Object> getTestTypeById(@PathVariable Long id) {
         if (!testTypesService.isExistsTestTypeById(id))
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
@@ -91,9 +100,18 @@ public class TestTypesController {
                 testTypesService.getTestTypeById(id),
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
-                statusCodeBundle.getViewTestTypeforIdSuccessMessage()));}
+                statusCodeBundle.getViewTestTypeforIdSuccessMessage()));
+    }
 
-    @GetMapping(EndpointURI.TESTTYPES_SEARCH)
+    @GetMapping(EndpointURI.TEST_TYPE_BY_PROJECT_ID)
+    public ResponseEntity<Object> getTestTypeByProjectId(@PathVariable Long id)
+    {
+        if (!projectService.existByProjectId(id)) return ResponseEntity.ok(new BaseResponse(RequestStatus.UNKNOWN.getStatus(), statusCodeBundle.getProjectNotExistCode(),statusCodeBundle.getProjectNotExistsMessage()));
+        else if (testTypesService.getTestTypesByProjectId(id).isEmpty()) return ResponseEntity.ok(new BaseResponse(RequestStatus.ERROR.getStatus(), statusCodeBundle.getFailureCode(),statusCodeBundle.getTestTypeNotMappedMessage()));
+        else return ResponseEntity.ok(new ContentResponse<>(Constants.TESTTYPES,testTypesService.getTestTypesByProjectId(id),statusCodeBundle.getCommonSuccessCode(),RequestStatus.SUCCESS.getStatus(),statusCodeBundle.getViewTestTypeByProjectIdSuccessMessage()));
+    }
+
+    @GetMapping(EndpointURI.TEST_TYPES_SEARCH)
     public ResponseEntity<Object> SearchTestTypesWithPagination(@RequestParam(name = "page") int page,
                                                                 @RequestParam(name = "size") int size,
                                                                 @RequestParam(name = "direction") String direction,
@@ -106,57 +124,63 @@ public class TestTypesController {
                 testTypesService.SearchTestTypesWithPagination(pageable, pagination, testTypesSearch),
                 RequestStatus.SUCCESS.getStatus(),
                 statusCodeBundle.getCommonSuccessCode(),
-                statusCodeBundle.getSuccessViewAllMessage()));}
+                statusCodeBundle.getSuccessViewAllMessageMainModules()));
+    }
 
-    @PostMapping(EndpointURI.TESTTYPE_IMPORT)
+    @PostMapping(EndpointURI.TEST_TYPE_IMPORT)
     public ResponseEntity<Object> importTestTypes(@RequestParam MultipartFile multipartFile) {
 
         Map<String, List<Integer>> errorMessages = new HashMap<>();
-        List<TestTypesRequest> testTypesRequestList;
+        Map<Integer,TestTypesRequest> testTypesRequestList;
+        Set<String> testTypeNames = new HashSet<>();
 
         try {
-            if (multipartFile.getOriginalFilename().endsWith(".csv")) {
+            if (!projectService.isCSVHeaderMatch(multipartFile) && !projectService.isExcelHeaderMatch(multipartFile)) {
+                return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(), statusCodeBundle.getFileFailureCode(), statusCodeBundle.getHeaderNotExistsMessage()));
+            }
+            if (Objects.requireNonNull(multipartFile.getOriginalFilename()).endsWith(".csv")) {
                 testTypesRequestList = testTypesService.csvProcess(multipartFile.getInputStream());
-
-            } else if (testTypesService.hasExcelFormat(multipartFile)) {
+            } else if (projectService.hasExcelFormat(multipartFile)) {
                 testTypesRequestList = testTypesService.excelProcess(multipartFile);
             } else {
                 return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(), statusCodeBundle.getFileFailureCode(), statusCodeBundle.getFileFailureMessage()));
             }
-
-            for (int rowIndex = 2; rowIndex <= testTypesRequestList.size() + 1; rowIndex++) {
-                TestTypesRequest testTypesRequest = testTypesRequestList.get(rowIndex - 2);
-
-                if (!Utils.isNotNullAndEmpty(testTypesRequest.getName())) {
-                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectNameEmptyMessage(), rowIndex);
+            for (Map.Entry<Integer, TestTypesRequest> entry : testTypesRequestList.entrySet()) {
+                if (!Utils.isNotNullAndEmpty(entry.getValue().getName())) {
+                    projectService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectNameEmptyMessage(), entry.getKey());
+                } else if (testTypeNames.contains(entry.getValue().getName())) {
+                    projectService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectNameDuplicateMessage(), entry.getKey());
+                } else {
+                    testTypeNames.add(entry.getValue().getName());
                 }
-                if (!Utils.isNotNullAndEmpty(testTypesRequest.getDescription())) {
-                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectDescriptionEmptyMessage(), rowIndex);
+                if (!Utils.isNotNullAndEmpty(entry.getValue().getDescription())) {
+                    projectService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectDescriptionEmptyMessage(), entry.getKey());
                 }
-                if (testTypesService.isExistsTestTypeByName(testTypesRequest.getName())) {
-                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getProjectNameAlReadyExistMessage(), rowIndex);
+                if (!Utils.isNotNullAndEmpty(entry.getValue().getDescription())) {
+                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getTestTypeDescriptionEmptyMessage(), entry.getKey());
+                }
+                if (testTypesService.isExistsTestTypeByName(entry.getValue().getName())) {
+                    testTypesService.addToErrorMessages(errorMessages, statusCodeBundle.getTestTypeNameAlReadyExistMessage(), entry.getKey());
                 }
             }
 
             if (!errorMessages.isEmpty()) {
                 return ResponseEntity.ok(new FileResponse(RequestStatus.FAILURE.getStatus(),
                         statusCodeBundle.getFailureCode(),
-                        statusCodeBundle.getBulkImportFailureMessage(),
+                        statusCodeBundle.getTestTypesNotSavedMessage(),
                         errorMessages));
             } else {
-                for (TestTypesRequest testTypesRequest : testTypesRequestList) {
-                    testTypesService.saveTestTypes(testTypesRequest);
+                for (Map.Entry<Integer, TestTypesRequest> entry : testTypesRequestList.entrySet()) {
+                    testTypesService.saveTestTypes(entry.getValue());
                 }
                 return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(),
                         statusCodeBundle.getCommonSuccessCode(),
-                        statusCodeBundle.getBulkImportMessage()));
+                        statusCodeBundle.getInsertTestTypesSuccessMessage()));
             }
         } catch (IOException e) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),
                     statusCodeBundle.getFailureCode(),
-                    statusCodeBundle.getBulkImportFailureMessage()));
+                    statusCodeBundle.getFileFailureMessage()));
         }
-
     }
-
 }
